Pushing command_namespace into build method
Prior to this commit, the method_missing allowed for override of the
command_namespace. However, this did not work for the
respond_to_missing as that is not something we can control calls to.

With this commit, I am saying the entire build for a given constant
should leverage the same series of command_namespaces.